fix 3

+++ LibraryApplication.Domain/Interfaces/Repositories/IUserRepository.cs
@@ -4,7 +4,7 @@ namespace LibraryApplication.Data.Interfaces.Repositories;
 
 public interface IUserRepository : IBaseCrudRepository<UserEntity>
 {
-    Task<int> UpdateUserBalance(int id, double amount);
+    Task<double> UpdateUserBalance(int id, double amount);
 
     Task<List<UserBalanceTransferEntity>> GetUserBalanceHistory(int id);
